Make default client event
